license: license under MIT
Previously this wasn't licensed - if we're moving it into
CollaboraOnline it's a good idea to give it a license. I've decided on
MIT as I wanted something permissive and MIT is used in nixpkgs already

I've also added reuse to keep track of the license and ensure that
any updates adhere to properly annotating files. That's in a new
"default" dev shell which I will use for things to manage this project.

Finally, I've added a github action which will use reuse to make sure
everything is up-to-snuff on new PRs